Scale correction for noise variance
For high-noise and low-occupancy settings, the targeted scale estimation suffered from over-estimation due to the finite noise variance being interpreted as contrast. There is now a correction which effectively compresses the scale estimate, so that the lowest confident primary scale estimate is redefined as 0 scale.

In the limit of large data this correction factor should vanish, and in a noisy and/or low-occupancy setting it will sooner under-estimate occupancy in the absence of evidence, as opposed to inferring noise as relevant scale.

As a consequence, regions with low confidence will automatically be redefined as negative scale and clipped to 0.

This also takes cae oc noise peak centering, I think.